+++ .github/workflows/deploy.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - master
-      - feature_deploy_github_actions
   pull_request:
     branches:
       - master
