+++ .husky/pre-commit
@@ -1,4 +1,4 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
-black blog simpleBlog
+black backend simpleBlog
 cd ./frontend && yarn prettier
