+++ backend/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class BlogConfig(AppConfig):
+class BackendConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "blog"
+    name = "backend"
