+++ simpleBlog/urls.py
@@ -12,7 +12,7 @@ Class-based views
     2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
 Including another URLconf
     1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+    2. Add a URL to urlpatterns:  path('backend/', include('backend.urls'))
 """
 
 from django.contrib import admin
@@ -20,7 +20,7 @@ from django.urls import path, include
 from graphene_django.views import GraphQLView
 
 urlpatterns = [
-    path("", include("blog.urls")),
+    path("", include("backend.urls")),
     path("graphql/", GraphQLView.as_view(graphiql=True)),
     path("admin/", admin.site.urls),
 ]
